types alocations
